Reconstruction of Stats system
- Plenty of bug fixes
- New Stat system
- Added tooltips
- Split stack changed from shift+LMB to MMB
- Added new stats

+++ Assets/Scripts/Player/PlayerUI.cs
@@ -32,14 +32,14 @@ public class PlayerUI : MonoBehaviour
         GetComponent<CharacterStats>().OnHealthChanged += OnHealthChanged;
     }
 
-    void OnHealthChanged(int maxHealth, int currentHealth)
+    void OnHealthChanged(float maxHealth, float currentHealth)
     {
         if (ui != null)
         {
             ui.gameObject.SetActive(true);
             lastMadeVisibleTime = Time.time;
 
-            float healthPercent = (float) currentHealth / maxHealth;
+            float healthPercent = currentHealth / maxHealth;
             healthSlider.fillAmount = healthPercent;
             if (currentHealth <= 0)
             {
